Removed controller on package name
