+++ deltaspike-projectstage/src/main/resources/META-INF/services/org.apache.deltaspike.core.spi.config.ConfigSource
@@ -1,1 +1,1 @@
-org.jboss.as.quickstarts.deltaspike.projectstage.controller.util.ProjectStageConfig
+org.jboss.as.quickstarts.deltaspike.projectstage.util.ProjectStageConfig
